+++ src/Fixtures/CastToClassWithStaticConstructor.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace EventSauce\ObjectHydrator\Fixtures;
 
 use Attribute;
-use EventSauce\ObjectHydrator\ObjectHydratorUsingReflection;
-use EventSauce\ObjectHydrator\ObjectHydrator;
+use EventSauce\ObjectHydrator\ObjectMapperUsingReflection;
+use EventSauce\ObjectHydrator\ObjectMapper;
 use EventSauce\ObjectHydrator\PropertyCaster;
 
 #[Attribute(Attribute::TARGET_PARAMETER)]
 final class CastToClassWithStaticConstructor implements PropertyCaster
 {
-    public function cast(mixed $value, ObjectHydrator $hydrator): mixed
+    public function cast(mixed $value, ObjectMapper $hydrator): mixed
     {
         return $hydrator->hydrateObject(ClassWithStaticConstructor::class, ['name' => $value]);
     }