+++ src/ObjectMapperCodeGeneratorSerializationTest.php
@@ -11,7 +11,7 @@ use function spl_object_hash;
 use function strpos;
 use function unlink;
 
-class DumpedObjectHydratorSerializationTest extends ObjectSerializerTestCase
+class ObjectMapperCodeGeneratorSerializationTest extends ObjectSerializationTestCase
 {
     private DefinitionProvider $defaultDefinitionProvider;
 
@@ -37,14 +37,14 @@ class DumpedObjectHydratorSerializationTest extends ObjectSerializerTestCase
         self::assertInstanceOf(ClassWithMappedStringProperty::class, $object);
     }
 
-    private function createDumpedObjectHydrator(string $directory, string $className, DefinitionProvider $definitionProvider): ObjectHydrator
+    private function createDumpedObjectHydrator(string $directory, string $className, DefinitionProvider $definitionProvider): ObjectMapper
     {
         if (class_exists($className, false)) {
             goto create_object_hydrator;
         }
 
         $classes = ConstructFinder::locatedIn($directory)->findClassNames();
-        $dumper = new ObjectHydratorDumper($definitionProvider);
+        $dumper = new ObjectMapperCodeGenerator($definitionProvider);
 
         $dumpedDefinition = $dumper->dump(
             $classes,
@@ -57,13 +57,13 @@ class DumpedObjectHydratorSerializationTest extends ObjectSerializerTestCase
         unlink($filename);
 
         create_object_hydrator:
-        /** @var ObjectHydrator $objectHydrator */
+        /** @var ObjectMapper $objectHydrator */
         $objectHydrator = new $className();
 
         return $objectHydrator;
     }
 
-    protected function createObjectHydrator(DefinitionProvider $definitionProvider = null): ObjectHydrator
+    protected function createObjectHydrator(DefinitionProvider $definitionProvider = null): ObjectMapper
     {
         $definitionProvider ??= $this->defaultDefinitionProvider;
         $className = 'AcmeCorp\\DumpedHydrator' . spl_object_hash($definitionProvider);
@@ -71,12 +71,12 @@ class DumpedObjectHydratorSerializationTest extends ObjectSerializerTestCase
         return $this->createDumpedObjectHydrator(__DIR__ . '/Fixtures', $className, $definitionProvider);
     }
 
-    public function objectHydrator(): ObjectHydrator
+    public function objectHydrator(): ObjectMapper
     {
         return $this->createDumpedObjectHydrator(__DIR__ . '/Fixtures', 'AcmeCorp\\DumpedHydrator', $this->defaultDefinitionProvider);
     }
 
-    protected function objectHydratorFor81(): ObjectHydrator
+    protected function objectHydratorFor81(): ObjectMapper
     {
         return $this->createDumpedObjectHydrator(__DIR__ . '/FixturesFor81', 'AcmeCorp\\DumpedHydratorFor81', $this->defaultDefinitionProvider);
     }